+++ MagicAPI/src/main/java/com/elmakers/mine/bukkit/api/magic/MageController.java
@@ -390,4 +390,9 @@ public interface MageController {
 
     Set<String> getMageClassKeys();
     MageClassTemplate getMageClassTemplate(String key);
+    
+    // LightAPI integration
+    boolean createLight(Location location, int lightlevel, boolean async);
+    boolean deleteLight(Location location, boolean async);
+    boolean updateLight(Location location);
 }